--- src/VirtoCommerce.CustomerExportImportModule.Core/Models/ImportContact.cs
@@ -1,9 +1,0 @@
-using VirtoCommerce.CustomerModule.Core.Model;
-
-namespace VirtoCommerce.CustomerExportImportModule.Core.Models
-{
-    public class ImportContact: ImportRecord<CsvContact>
-    {
-        public Contact Contact { get; set; }
-    }
-}
