+++ src/VirtoCommerce.CustomerExportImportModule.Data/Services/CsvPagedOrganizationDataImporter.cs
@@ -12,20 +12,20 @@ using VirtoCommerce.Platform.Core.Common;
 
 namespace VirtoCommerce.CustomerExportImportModule.Data.Services
 {
-    public sealed class CsvPagedOrganizationDataImporter : CsvPagedDataImporter<CsvOrganization, Organization>
+    public sealed class CsvPagedOrganizationDataImporter : CsvPagedDataImporter<ImportableOrganization, Organization>
     {
         private readonly IMemberService _memberService;
 
         public override string MemberType => nameof(Organization);
 
-        public CsvPagedOrganizationDataImporter(IMemberService memberService, IMemberSearchService memberSearchService, ICsvCustomerDataValidator dataValidator, IValidator<ImportRecord<CsvOrganization>[]> importOrganizationValidator
+        public CsvPagedOrganizationDataImporter(IMemberService memberService, IMemberSearchService memberSearchService, ICsvCustomerDataValidator dataValidator, IValidator<ImportRecord<ImportableOrganization>[]> importOrganizationValidator
             , ICustomerImportPagedDataSourceFactory dataSourceFactory, ICsvCustomerImportReporterFactory importReporterFactory, IBlobUrlResolver blobUrlResolver)
         : base(memberSearchService, dataValidator, dataSourceFactory, importOrganizationValidator, importReporterFactory, blobUrlResolver)
         {
             _memberService = memberService;
         }
 
-        protected override async Task ProcessChunkAsync(ImportDataRequest request, Action<ImportProgressInfo> progressCallback, ICustomerImportPagedDataSource<CsvOrganization> dataSource,
+        protected override async Task ProcessChunkAsync(ImportDataRequest request, Action<ImportProgressInfo> progressCallback, ICustomerImportPagedDataSource<ImportableOrganization> dataSource,
             ImportErrorsContext errorsContext, ImportProgressInfo importProgress, ICsvCustomerImportReporter importReporter)
         {
             var importOrganizations = dataSource.Items
@@ -52,7 +52,7 @@ namespace VirtoCommerce.CustomerExportImportModule.Data.Services
                 var validationResult = await ValidateAsync(importOrganizations, importReporter);
 
                 var invalidImportOrganizations = validationResult.Errors
-                    .Select(x => (x.CustomState as ImportValidationState<CsvOrganization>)?.InvalidRecord).Distinct().ToArray();
+                    .Select(x => (x.CustomState as ImportValidationState<ImportableOrganization>)?.InvalidRecord).Distinct().ToArray();
 
                 importOrganizations = importOrganizations.Except(invalidImportOrganizations).ToArray();
 
@@ -92,24 +92,24 @@ namespace VirtoCommerce.CustomerExportImportModule.Data.Services
         }
 
 
-        private static void PatchExistedOrganizations(IEnumerable<Organization> existedOrganizations, ImportRecord<CsvOrganization>[] updateImportOrganizations)
+        private static void PatchExistedOrganizations(IEnumerable<Organization> existedOrganizations, ImportRecord<ImportableOrganization>[] updateImportOrganizations)
         {
             foreach (var existedOrganization in existedOrganizations)
             {
                 var importOrganization = updateImportOrganizations.LastOrDefault(x => existedOrganization.Id.EqualsInvariant(x.Record.Id)
                                                                             || (!existedOrganization.OuterId.IsNullOrEmpty() && existedOrganization.OuterId.EqualsInvariant(x.Record.OuterId)));
 
-                importOrganization?.Record.PatchOrganization(existedOrganization);
+                importOrganization?.Record.PatchModel(existedOrganization);
             }
         }
 
-        private static Organization[] CreateNewOrganizations(ImportRecord<CsvOrganization>[] createImportOrganizations)
+        private static Organization[] CreateNewOrganizations(ImportRecord<ImportableOrganization>[] createImportOrganizations)
         {
             var newOrganizations = createImportOrganizations.Select(x =>
             {
                 var organization = AbstractTypeFactory<Organization>.TryCreateInstance();
 
-                x.Record.PatchOrganization(organization);
+                x.Record.PatchModel(organization);
 
                 return organization;
             }).ToArray();
