+++ src/VirtoCommerce.CustomerExportImportModule.Data/Validation/ImportContactsValidator.cs
@@ -7,7 +7,7 @@ using VirtoCommerce.Platform.Core.Security;
 
 namespace VirtoCommerce.CustomerExportImportModule.Data.Validation
 {
-    public sealed class ImportContactsValidator: AbstractValidator<ImportRecord<CsvContact>[]>
+    public sealed class ImportContactsValidator: AbstractValidator<ImportRecord<ImportableContact>[]>
     {
         private readonly ICountriesService _countriesService;
         private readonly IDynamicPropertyDictionaryItemsSearchService _dynamicPropertyDictionaryItemsSearchService;
@@ -24,8 +24,8 @@ namespace VirtoCommerce.CustomerExportImportModule.Data.Validation
 
         private void AttachValidators()
         {
-            RuleFor(importRecords => importRecords).SetValidator(_ => new ImportEntitiesAreNotDuplicatesValidator<CsvContact>());
-            RuleForEach(importRecords => importRecords).SetValidator(new ImportMemberValidator<CsvContact>(_countriesService, _dynamicPropertyDictionaryItemsSearchService));
+            RuleFor(importRecords => importRecords).SetValidator(_ => new ImportEntitiesAreNotDuplicatesValidator<ImportableContact>());
+            RuleForEach(importRecords => importRecords).SetValidator(new ImportMemberValidator<ImportableContact>(_countriesService, _dynamicPropertyDictionaryItemsSearchService));
             RuleForEach(importRecords => importRecords).SetValidator(new ImportContactValidator(_signInManager.UserManager));
         }
     }
