+++ tests/VirtoCommerce.CustomerExportImportModule.Tests/CustomerImportPagedDataSourceTests.cs
@@ -34,7 +34,7 @@ namespace VirtoCommerce.CustomerExportImportModule.Tests
             var dynamicPropertySearchService = TestHelper.GetDynamicPropertySearchService(DynamicProperties);
             var dynamicPropertyDictionaryItemsSearchService = TestHelper.GetDynamicPropertyDictionaryItemsSearchService(DynamicProperties, new Dictionary<string, IList<DynamicPropertyDictionaryItem>>());
             var customerImportPagedDataSourceFactory = TestHelper.GetCustomerImportPagedDataSourceFactory(blobStorageProvider, dynamicPropertySearchService, dynamicPropertyDictionaryItemsSearchService);
-            using var customerImportPagedDataSource = await customerImportPagedDataSourceFactory.CreateAsync<CsvContact, Contact>(CsvFileName, 10);
+            using var customerImportPagedDataSource = await customerImportPagedDataSourceFactory.CreateAsync<ImportableContact, Contact>(CsvFileName, 10);
 
             // Act
             var totalCount = customerImportPagedDataSource.GetTotalCount();
@@ -52,7 +52,7 @@ namespace VirtoCommerce.CustomerExportImportModule.Tests
             var dynamicPropertySearchService = TestHelper.GetDynamicPropertySearchService(DynamicProperties);
             var dynamicPropertyDictionaryItemsSearchService = TestHelper.GetDynamicPropertyDictionaryItemsSearchService(DynamicProperties, new Dictionary<string, IList<DynamicPropertyDictionaryItem>>());
             var customerImportPagedDataSourceFactory = TestHelper.GetCustomerImportPagedDataSourceFactory(blobStorageProvider, dynamicPropertySearchService, dynamicPropertyDictionaryItemsSearchService);
-            using var customerImportPagedDataSource = await customerImportPagedDataSourceFactory.CreateAsync<CsvContact, Contact>(CsvFileName, 10);
+            using var customerImportPagedDataSource = await customerImportPagedDataSourceFactory.CreateAsync<ImportableContact, Contact>(CsvFileName, 10);
 
             // Act
             customerImportPagedDataSource.GetTotalCount();
@@ -79,7 +79,7 @@ namespace VirtoCommerce.CustomerExportImportModule.Tests
                 var dynamicPropertySearchService = TestHelper.GetDynamicPropertySearchService(DynamicProperties);
                 var dynamicPropertyDictionaryItemsSearchService = TestHelper.GetDynamicPropertyDictionaryItemsSearchService(DynamicProperties, new Dictionary<string, IList<DynamicPropertyDictionaryItem>>());
                 var customerImportPagedDataSourceFactory = TestHelper.GetCustomerImportPagedDataSourceFactory(blobStorageProvider, dynamicPropertySearchService, dynamicPropertyDictionaryItemsSearchService);
-                using var customerImportPagedDataSource = await customerImportPagedDataSourceFactory.CreateAsync<CsvContact, Contact>(CsvFileName, 10);
+                using var customerImportPagedDataSource = await customerImportPagedDataSourceFactory.CreateAsync<ImportableContact, Contact>(CsvFileName, 10);
 
                 // Act
                 await customerImportPagedDataSource.FetchAsync();
@@ -98,7 +98,7 @@ namespace VirtoCommerce.CustomerExportImportModule.Tests
             var dynamicPropertySearchService = TestHelper.GetDynamicPropertySearchService(DynamicProperties);
             var dynamicPropertyDictionaryItemsSearchService = TestHelper.GetDynamicPropertyDictionaryItemsSearchService(DynamicProperties, new Dictionary<string, IList<DynamicPropertyDictionaryItem>>());
             var customerImportPagedDataSourceFactory = TestHelper.GetCustomerImportPagedDataSourceFactory(blobStorageProvider, dynamicPropertySearchService, dynamicPropertyDictionaryItemsSearchService);
-            using var customerImportPagedDataSource = await customerImportPagedDataSourceFactory.CreateAsync<CsvContact, Contact>(CsvFileName, 1);
+            using var customerImportPagedDataSource = await customerImportPagedDataSourceFactory.CreateAsync<ImportableContact, Contact>(CsvFileName, 1);
 
             // Act
             await customerImportPagedDataSource.FetchAsync();
@@ -117,7 +117,7 @@ namespace VirtoCommerce.CustomerExportImportModule.Tests
             var dynamicPropertySearchService = TestHelper.GetDynamicPropertySearchService(DynamicProperties);
             var dynamicPropertyDictionaryItemsSearchService = TestHelper.GetDynamicPropertyDictionaryItemsSearchService(DynamicProperties, new Dictionary<string, IList<DynamicPropertyDictionaryItem>>());
             var customerImportPagedDataSourceFactory = TestHelper.GetCustomerImportPagedDataSourceFactory(blobStorageProvider, dynamicPropertySearchService, dynamicPropertyDictionaryItemsSearchService);
-            using var customerImportPagedDataSource = await customerImportPagedDataSourceFactory.CreateAsync<CsvContact, Contact>(CsvFileName, 1);
+            using var customerImportPagedDataSource = await customerImportPagedDataSourceFactory.CreateAsync<ImportableContact, Contact>(CsvFileName, 1);
 
             // Act
             await customerImportPagedDataSource.FetchAsync();
@@ -137,7 +137,7 @@ namespace VirtoCommerce.CustomerExportImportModule.Tests
             var dynamicPropertySearchService = TestHelper.GetDynamicPropertySearchService(DynamicProperties);
             var dynamicPropertyDictionaryItemsSearchService = TestHelper.GetDynamicPropertyDictionaryItemsSearchService(DynamicProperties, new Dictionary<string, IList<DynamicPropertyDictionaryItem>>());
             var customerImportPagedDataSourceFactory = TestHelper.GetCustomerImportPagedDataSourceFactory(blobStorageProvider, dynamicPropertySearchService, dynamicPropertyDictionaryItemsSearchService);
-            using var customerImportPagedDataSource = await customerImportPagedDataSourceFactory.CreateAsync<CsvContact, Contact>(CsvFileName, 1);
+            using var customerImportPagedDataSource = await customerImportPagedDataSourceFactory.CreateAsync<ImportableContact, Contact>(CsvFileName, 1);
 
             // Act
             var result = await customerImportPagedDataSource.FetchAsync();
@@ -155,7 +155,7 @@ namespace VirtoCommerce.CustomerExportImportModule.Tests
             var dynamicPropertySearchService = TestHelper.GetDynamicPropertySearchService(DynamicProperties);
             var dynamicPropertyDictionaryItemsSearchService = TestHelper.GetDynamicPropertyDictionaryItemsSearchService(DynamicProperties, new Dictionary<string, IList<DynamicPropertyDictionaryItem>>());
             var customerImportPagedDataSourceFactory = TestHelper.GetCustomerImportPagedDataSourceFactory(blobStorageProvider, dynamicPropertySearchService, dynamicPropertyDictionaryItemsSearchService);
-            using var customerImportPagedDataSource = await customerImportPagedDataSourceFactory.CreateAsync<CsvContact, Contact>(CsvFileName, 10);
+            using var customerImportPagedDataSource = await customerImportPagedDataSourceFactory.CreateAsync<ImportableContact, Contact>(CsvFileName, 10);
 
             // Act
             await customerImportPagedDataSource.FetchAsync();
@@ -174,7 +174,7 @@ namespace VirtoCommerce.CustomerExportImportModule.Tests
             var dynamicPropertySearchService = TestHelper.GetDynamicPropertySearchService(DynamicProperties);
             var dynamicPropertyDictionaryItemsSearchService = TestHelper.GetDynamicPropertyDictionaryItemsSearchService(DynamicProperties, new Dictionary<string, IList<DynamicPropertyDictionaryItem>>());
             var customerImportPagedDataSourceFactory = TestHelper.GetCustomerImportPagedDataSourceFactory(blobStorageProvider, dynamicPropertySearchService, dynamicPropertyDictionaryItemsSearchService);
-            using var customerImportPagedDataSource = await customerImportPagedDataSourceFactory.CreateAsync<CsvContact, Contact>(CsvFileName, 10);
+            using var customerImportPagedDataSource = await customerImportPagedDataSourceFactory.CreateAsync<ImportableContact, Contact>(CsvFileName, 10);
 
             // Act
             await customerImportPagedDataSource.FetchAsync();
